requirements.txt
twitchio==2.8.2
flask==3.0.3
flask-mysqldb==1.0.1
pygame==2.6.0
gTTS==2.5.1
pydub==0.25.1
librosa==0.10.2.post1
soundfile==0.12.1
numpy==1.26.4
requests==2.32.3
pygetwindow==0.0.9
unidecode==1.3.8
obs-websocket-py==1.5.1
python-dotenv==1.0.1
playsound==1.3.0
pywin32==306
